Set timezone for Rails
Nova Labs happens to be on the east coast, so lets set the timezone
there, this saves us the extra work for saving timezones for each event
or user.

Considered setting the `timezone` on the event row for each event, but
then I would also have to convert other dates (like when a user was
updated), so figured the one-line change is worth a go. May never have
to revisit!