Default admin to false on sync
When syncing a user that has had the `[nlgroup] wautils` sign-off
removed it was not unsetting the `admin` flag on the user. This resulted
in people who should not have access to edit sign offs having that
ability.

Thanks to @sudobob for a clear bug report that let me discover and
recreate the problem.