Default google workspace settings to users attrs
This will set the name and email address in google based on the users'
first and last name. It is not required, but will populate the form with
the conventions.

The account in Google is not associated with the WA record in any way,
but we make it a nested resource to ease the setup process.

Removes the telephone number from the integromat request.

Fixes a bug in clear search link on users index.