Verify the WA ID is the same in the session
There was bug reports that folks were logged in as the wrong person. The
only way I can imagine this happens is with the sync process running and
the underlying `user_id` changes and a different member is assigned to
that column.

This may be because we dropped the DB at one point to re-sync other
fields, but adding the check will prevent this from happening in the
future.