OTP support for Formbot
Formbot is removing the need for a password! The changes below are to
support linking directly to an event with a OTP as a token in the URL
that will give an instructor the ability to submit the class to the
finance folks.

Ths OTP uses a TOTP approach with a shared secret between Watto and
Formbot. For reference I made a short script to test the approach with
the [same library](https://github.com/yeojz/otplib) that Formbot will use:

```javascript

// Usage: auth.mjs BASE32SECRET CODE

import { authenticator } from 'otplib';
const secret = process.argv[2];
const token = process.argv[3];
const check = authenticator.check(token, secret);
const verify = authenticator.verify({ token, secret });

console.log('token: ', token, ' check: ', check, ' verify: ', verify);
```

Since the OTP is short lived all links to Formbot will be routed through
a new controller that will add `Event#uid` and the newly generated code
to the url and redirect to the other host.

For deployment coordination there is a config option set to use the
legacy URL format:

```
FORMBOT_USE_LEGACY_FORMAT=true
```

Once the other service is updated that var will be deleted from the
environment variables and the wautils service restarted.