Allow user sync with missing field
When fields are added the user sync would fail when trying to set user
field values. Most of the time we don't care about the value of those
fields (and they'll be updated next sync AFTER the fields are added).

In those cases log there was a problem and keep going with the sync
(which is typically because of a webhook)