Update class names to match production
Sorted the list to minimize future churn

Fixed up a few names to match production (including adding a double
space to one of the names)

Dropped one tool that doesn't exist.

Added a bang (!) to the `find_by` to throw a more useful exception if
this gets out of sync again in the future. And it will.