Sort sign offs by position
Wild Apricot has a position attribute for an allowed value. This was
already being stored in the DB, so lets update the sort to use the,
well, sort order.

This will allow the org to manage how things are displayed in the
sign off pages with the WA admin portal. So more naming things with
prefixes just to get things sorted how they want.