Update the Fields model to match WA
Primarily this normalizes the Field User Value to Field Allowed Value,
so the field label is pulled out of the Allowed table. This was done
because a sync wouldn't update all the labels if someone re-named the
allowed value in the WA portal.

In other words: Changing "[equipment] WWY_Various_Drill Press" to
"[WW-Y] Drill Press" would not also update the list of user's signoffs.
This way we update in one place (the filed_allowed_values table) and all
the labels update.