Usability changes for instructors
Did a little ad-hod hallway testing with one of the instructors and got
the feedback "it takes 12 clicks to do what I used to be able to do in
2" so wanted address that.

Seems like the biggest point of confusion was the usage of the search
and past events. To address this I did two things:

On the welcome page, it now shows past events going back 12 days.
Hopefully this will make it so there is one click after logging in to
get to the right event.

On the event index page I added a button to the search field. Having to
hit enter was confusing with the combination of the "past events"
checkbox. Hopefully this provides a little affordance to give clues to
how the form wants to work.