Filter out canceled event registrations
Fixes #27

Instructors found it confusing to have canceled registrations showing up
in the list of attendees with looking at the event detail page. This
change filters those out of the list.

Turns out the only way to have an canceled registration is when we get a
webhook letting the system know we need to cancel them. Normally the API
wouldn't return any in the list when we fetch event details.