Add page to edit the badge number
I debated if we should make a user fields controller and have a generic
way to edit any fields (like we do in the admin pages) but opted not to.
I think going forward we may need to edit multiple user fields. That
will require some refactor for the WAAPI lib, but seems doable.

For not edit user will:

- POST changes to the WA API
- Pull the updated user attributes
- Update the user in the DB
- Redirect to the show user page