Force sign off classes to the top of the waitlist
The waitlist page uses the order of the hash to determine how it is
displayed on the page that list the different class types. If the
support desk folks reorder the google sheet then project classes might be at
the top of the page.

The majority of members care about getting on the waitlist for sign off
classes the most -- or at least that is where they start. This will
rebuild the hash so that ruby orders the keys with `SIGN_OFF_CLASS`
first.