Expose badge number and door group for all members
This was needed for coordinating with Robotics and folks not being able
to see what the access should be.

It was decided that exposing this information to members was worth it
because making it viewable might be that all the
stewards/leaders/volunteers could see the badge and access level.

See slack for details https://nova-labs-org.slack.com/archives/CE0T0ULP7/p1763828194039209